+++ pytorch_lightning/metrics/regression/__init__.py
@@ -1,1 +1,4 @@
-from pytorch_lightning.metrics.regression.mean_error import MeanSquaredError, MeanAbsoluteError, MeanSquaredLogError
+from pytorch_lightning.metrics.regression.mean_squared_error import MeanSquaredError
+from pytorch_lightning.metrics.regression.mean_absolute_error import MeanAbsoluteError
+from pytorch_lightning.metrics.regression.mean_squared_log_error import MeanSquaredLogError
+
